pyproject.toml
[project]
name = "histological-image-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "huggingface-hub>=0.20",
    "matplotlib>=3.5",
    "nibabel>=5.0",
    "numpy>=1.24",
    "Pillow>=10.0",
    "pynrrd>=1.1.0",
    "scipy>=1.10",
    "svgpathtools>=1.6",
    "torch>=2.0",
    "torchvision>=0.15",
    "tqdm>=4.66",
    "transformers>=4.40",
]

[project.optional-dependencies]
exploration = [
    "allensdk>=0.16.3",
    "fiftyone>=1.13.3",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.5",
    "kagglehub>=1.0.0",
    "pydicom==2.4.4",
    "requests>=2.32.5",
    "rt-utils>=1.2.7",
]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/histological_image_analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
